Apply kNN to MultiSearch's child SearchRequest
+ This change also drops a couple files, and groups like content into
  existing files.
+ With this multi partition search and gRPC scatter gather will work
  as expected.
