+++ index_alias_impl.go
@@ -430,21 +430,7 @@ func (i *indexAliasImpl) Swap(in, out []Index) {
 // Perhaps that part needs to be optional,
 // could be slower in remote usages.
 func createChildSearchRequest(req *SearchRequest) *SearchRequest {
-	rv := SearchRequest{
-		Query:            req.Query,
-		Size:             req.Size + req.From,
-		From:             0,
-		Highlight:        req.Highlight,
-		Fields:           req.Fields,
-		Facets:           req.Facets,
-		Explain:          req.Explain,
-		Sort:             req.Sort.Copy(),
-		IncludeLocations: req.IncludeLocations,
-		Score:            req.Score,
-		SearchAfter:      req.SearchAfter,
-		SearchBefore:     req.SearchBefore,
-	}
-	return &rv
+	return copySearchRequest(req)
 }
 
 type asyncSearchResult struct {
@@ -467,10 +453,6 @@ func MultiSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*Se
 		req.SearchAfter = req.SearchBefore
 		req.SearchBefore = nil
 	}
-
-	// Using a disjunction query to get union of results from KNN query
-	// and the original query
-	req.Query = disjunctQueryWithKNN(req)
 
 	// run search on each index in separate go routine
 	var waitGroup sync.WaitGroup
