+++ search_no_knn.go
@@ -123,3 +123,27 @@ func (r *SearchRequest) UnmarshalJSON(input []byte) error {
 	return nil
 
 }
+
+// -----------------------------------------------------------------------------
+
+func copySearchRequest(req *SearchRequest) *SearchRequest {
+	rv := SearchRequest{
+		Query:            req.Query,
+		Size:             req.Size + req.From,
+		From:             0,
+		Highlight:        req.Highlight,
+		Fields:           req.Fields,
+		Facets:           req.Facets,
+		Explain:          req.Explain,
+		Sort:             req.Sort.Copy(),
+		IncludeLocations: req.IncludeLocations,
+		Score:            req.Score,
+		SearchAfter:      req.SearchAfter,
+		SearchBefore:     req.SearchBefore,
+	}
+	return &rv
+}
+
+func disjunctQueryWithKNN(req *SearchRequest) query.Query {
+	return req.Query
+}
